MAINT: move _support_alternative_backends from special to _lib
