+++ scipy/special/_support_alternative_backends.py
@@ -1,11 +1,7 @@
-import os
 import sys
-import functools
 
 import numpy as np
-from scipy._lib._array_api import (
-    array_namespace, scipy_namespace_for, is_numpy
-)
+from scipy._lib._array_api import support_alternative_backends, SCIPY_ARRAY_API
 from . import _ufuncs
 # These don't really need to be imported, but otherwise IDEs might not realize
 # that these are defined in this file / report an error in __init__.py
@@ -14,38 +10,6 @@ from ._ufuncs import (
     gammainc, gammaincc, logit, expit, entr, rel_entr, xlogy,  # noqa: F401
     chdtr, chdtrc, betainc, betaincc, stdtr  # noqa: F401
 )
-
-_SCIPY_ARRAY_API = os.environ.get("SCIPY_ARRAY_API", False)
-array_api_compat_prefix = "scipy._lib.array_api_compat"
-
-
-def get_array_special_func(f_name, xp, n_array_args):
-    spx = scipy_namespace_for(xp)
-    f = None
-    if is_numpy(xp):
-        f = getattr(_ufuncs, f_name, None)
-    elif spx is not None:
-        f = getattr(spx.special, f_name, None)
-
-    if f is not None:
-        return f
-
-    # if generic array-API implementation is available, use that;
-    # otherwise, fall back to NumPy/SciPy
-    if f_name in _generic_implementations and spx is not None:
-        _f = _generic_implementations[f_name](xp=xp, spx=spx)
-        if _f is not None:
-            return _f
-
-    _f = getattr(_ufuncs, f_name, None)
-    def f(*args, _f=_f, _xp=xp, **kwargs):
-        array_args = args[:n_array_args]
-        other_args = args[n_array_args:]
-        array_args = [np.asarray(arg) for arg in array_args]
-        out = _f(*array_args, *other_args, **kwargs)
-        return _xp.asarray(out)
-
-    return f
 
 
 def _get_shape_dtype(*args, xp):
@@ -79,6 +43,8 @@ def _xlogy(xp, spx):
 
 
 def _chdtr(xp, spx):
+    if spx is None:
+        return None
     # The difference between this and just using `gammainc`
     # defined by `get_array_special_func` is that if `gammainc`
     # isn't found, we don't want to use the SciPy version; we'll
@@ -98,6 +64,8 @@ def _chdtr(xp, spx):
 
 
 def _chdtrc(xp, spx):
+    if spx is None:
+        return None
     # The difference between this and just using `gammaincc`
     # defined by `get_array_special_func` is that if `gammaincc`
     # isn't found, we don't want to use the SciPy version; we'll
@@ -117,6 +85,8 @@ def _chdtrc(xp, spx):
 
 
 def _betaincc(xp, spx):
+    if spx is None:
+        return None
     betainc = getattr(spx.special, 'betainc', None)  # noqa: F811
     if betainc is None and hasattr(xp, 'special'):
         betainc = getattr(xp.special, 'betainc', None)
@@ -130,6 +100,8 @@ def _betaincc(xp, spx):
 
 
 def _stdtr(xp, spx):
+    if spx is None:
+        return None
     betainc = getattr(spx.special, 'betainc', None)  # noqa: F811
     if betainc is None and hasattr(xp, 'special'):
         betainc = getattr(xp.special, 'betainc', None)
@@ -151,21 +123,6 @@ _generic_implementations = {'rel_entr': _rel_entr,
                             'betaincc': _betaincc,
                             'stdtr': _stdtr,
                             }
-
-
-# functools.wraps doesn't work because:
-# 'numpy.ufunc' object has no attribute '__module__'
-def support_alternative_backends(f_name, n_array_args):
-    func = getattr(_ufuncs, f_name)
-
-    @functools.wraps(func)
-    def wrapped(*args, **kwargs):
-        xp = array_namespace(*args[:n_array_args])
-        f = get_array_special_func(f_name, xp, n_array_args)
-        return f(*args, **kwargs)
-
-    return wrapped
-
 
 array_special_func_map = {
     'log_ndtr': 1,
@@ -193,7 +150,8 @@ array_special_func_map = {
 }
 
 for f_name, n_array_args in array_special_func_map.items():
-    f = (support_alternative_backends(f_name, n_array_args) if _SCIPY_ARRAY_API
+    f = (support_alternative_backends(f_name, n_array_args, _ufuncs, 'special',
+                                      _generic_implementations) if SCIPY_ARRAY_API
          else getattr(_ufuncs, f_name))
     sys.modules[__name__].__dict__[f_name] = f
